+++ packages/cosmos-ts/src/types.ts
@@ -8,6 +8,7 @@ export enum CosmosChainId {
   Injective = 'injective-1',
   Cosmoshub = 'cosmoshub-4',
   Terra = 'columbus-5',
+  TerraUST = 'columbus-5',
 }
 
 export enum TestnetCosmosChainId {
@@ -20,10 +21,16 @@ export const cosmosNativeDenomsFromChainId = {
     ...Erc20TokenMeta.getMetaBySymbol('ATOM'),
     denom: 'uatom',
   },
-  [CosmosChainId.Terra]: {
-    ...Erc20TokenMeta.getMetaBySymbol('LUNA'),
-    denom: 'uluna',
-  },
+  [CosmosChainId.Terra]: [
+    {
+      ...Erc20TokenMeta.getMetaBySymbol('LUNA'),
+      denom: 'uluna',
+    },
+    {
+      ...Erc20TokenMeta.getMetaBySymbol('UST'),
+      denom: 'uusd',
+    },
+  ],
   [CosmosChainId.Injective]: {
     ...Erc20TokenMeta.getMetaBySymbol('INJ'),
     denom: 'inj',
@@ -36,4 +43,4 @@ export const cosmosNativeDenomsFromChainId = {
     ...Erc20TokenMeta.getMetaBySymbol('INJ'),
     denom: 'inj',
   },
-} as Record<string, TokenMetaWithNativeDenom>
+} as Record<string, TokenMetaWithNativeDenom | TokenMetaWithNativeDenom[]>