+++ packages/spot-consumer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.84](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.83...@injectivelabs/spot-consumer@0.0.84) (2021-12-05)
+
+
+### Bug Fixes
+
+* cosmos ts + packages bump ([8de8ab2](https://github.com/InjectiveLabs/injective-ts/commit/8de8ab230385579995566634c5c1fc1a97982c49))
+
+
+
+
+
 ## [0.0.83](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.80...@injectivelabs/spot-consumer@0.0.83) (2021-12-03)
 
 **Note:** Version bump only for package @injectivelabs/spot-consumer